+++ tests/plugins/html/function.hello.php
@@ -1,0 +1,6 @@
+<?php
+
+function jtpl_function_html_hello(\Jelix\Castor\CastorCore $tpl)
+{
+    echo "<p>Hello world!</p>";
+}