+++ tests/plugins/text/function.hello.php
@@ -1,0 +1,6 @@
+<?php
+
+function jtpl_function_text_hello(\Jelix\Castor\CastorCore $tpl)
+{
+    echo "Hello world!\n";
+}